+++ HISTORY.rst
@@ -3,6 +3,13 @@
 Release History
 ---------------
 
+Upcoming
+++++++++
+
+- Added support for Box Developer Edition. This includes JWT auth (auth as enterprise or as app user),
+  and `create_user` functionality.
+
+
 1.1.7 (2015-05-28)
 ++++++++++++++++++